북 작업
